pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rl-dino-agent"
version = "0.1.0"
description = "Visual-input RL agent for the RL Dino browser game."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "gymnasium>=0.29,<1.1",
  "matplotlib>=3.8,<4.0",
  "numpy>=1.26,<3.0",
  "opencv-python>=4.10,<5.0",
  "playwright>=1.52,<2.0",
  "PyYAML>=6.0,<7.0",
  "rich>=13.7,<14.0",
  "stable-baselines3>=2.3,<3.0",
  "tensorboard>=2.16,<3.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
